remove stlyes file from screen dir
